Move causal snapshot wire code to the right submodule
